requirements.txt
mypy
torch
Pillow
git+https://github.com/PramaLLC/BEN2.git#egg=ben2
opencv-python
